updated food tag
