+++ src/Application/Common/Interfaces/IFoodTagsRepository.cs
@@ -5,6 +5,5 @@ namespace Application.Common.Interfaces
     public interface IFoodTagsRepository
     {
         public IEnumerable<Domain.Entities.FoodTag> GetAll();
-        public IEnumerable<Domain.Entities.FoodTag> Filter(IEnumerable<Domain.Entities.NutrientItem> nutrients);
     }
 }
